Address comments from @ikhoon
